requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
tensorflow>=2.18.0
opencv-python-headless==4.10.0.84
Pillow>=10.2.0
numpy>=1.26.0,<2.0.0
scikit-learn==1.3.2
gunicorn==21.2.0
